Intersections refactoring.
